Add local mode back and minor bug fix
Force the nearbottom behavior to return the plugin's return value, and then use that to implement a local mode like the docs say should exist.

Use this by including behaviors/local.js in your script and then specify behavior:'local' in your options.

Additionally, you need to specify what the container is that will be scrolling (this defaults to the window).  I used options.binder for this, i.e. binder:'#myscrollingdiv'