Force debug option to be stored as bool rather than string.
This will output (via `wp_localize_script()`) as either "1" or "0" (rather than "true" or "" as did previosly).

When javascript evaluates the truthiness of a string "1" evaluates to true, and "0" (or "") evaluates to false.

This saves the need to manually check for the string "true" on each frontend load.

Additionally, on the options screen, use core's `checked()` API to toggle the checkbox, rather than custom coding a check.