Upgrade database sequentially
Move the 2.6 upgrade bath below the 2.5 upgrade path, and do not make the if statement an else.

This allows a user to upgrade from pre-2.5 to 2.6 (previously if a user was < 2.6, they would not trigger the 2.5 upgrade) and allows upgrades down the line (e.g, 2.7) to simply upgrade from 2.6 (because it can assume the 2.5 upgrade was triggered).