Fix for url format /page/2
When Wordpress permalinks are set to be custom /%category%/%postname%/,
this breaks the infinite scroll because it is unable to find the page
number as the url is represented as category/page/2 instead of
category?page=2.

The following code changes fix this issue while not impacting any
existing functionality.

The change is to the '=' sign in the regular expression. This needs to
be changed to be either an '=' OR a '/' as is used in wordpress.
Thus the '=' in the following path matching lines becomes [=\/].